requirement.txt
alembic==1.13.0
amqp==5.3.1
annotated-types==0.7.0
ansi2html==1.9.2
asgiref==3.9.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.5
billiard==4.2.2
black==23.11.0
blinker==1.9.0
blis==0.7.11
branca==0.8.1
catalogue==2.0.10
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cligj==0.7.2
cloudpathlib==0.16.0
comm==0.2.3
confection==0.1.5
constantly==23.10.4
contourpy==1.3.2
coverage==7.10.5
cryptography==45.0.6
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.16
decorator==5.2.1
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fake-useragent==1.4.0
filelock==3.19.1
fiona==1.10.1
flake8==6.1.0
Flask==3.0.3
folium==0.15.0
fonttools==4.59.2
fsspec==2025.9.0
geopandas==0.14.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
itemadapter==0.12.1
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.1
Mako==1.3.10
marisa-trie==1.3.1
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
murmurhash==1.0.13
mypy==1.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==1.26.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.4
parsel==1.10.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.26
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
plotly==5.17.0
pluggy==1.6.0
preshed==3.0.10
prometheus-client==0.19.0
prompt_toolkit==3.0.52
Protego==0.5.0
psutil==7.0.0
psycopg==3.2.10
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==13.0.0
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyproj==3.7.1
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.0.2
queuelib==1.8.0
redis==6.4.0
regex==2025.9.18
requests==2.31.0
requests-file==2.1.0
retrying==1.4.2
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
Scrapy==2.11.0
scrapy-playwright==0.0.44
scrapy-redis==0.7.2
scrapy-splash==0.8.0
seaborn==0.13.2
selenium==4.35.0
sentence-transformers==2.2.2
sentencepiece==0.2.1
sentry-sdk==1.38.0
service-identity==24.2.0
shapely==2.0.2
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
thinc==8.2.5
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
torchvision==0.16.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
trio==0.30.0
trio-websocket==0.12.2
Twisted==22.10.0
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver==3.5.5
Unidecode==1.3.7
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
whitenoise==6.6.0
wsproto==1.2.0
zipp==3.23.0
zope.interface==7.2
